Update and rename moonpig-data-breach-2015 to moonpig-data-breach-2015.md
Does Moonpig know this is going up on the site? Because even if the issue was completely resolved and made Moonpig a safer company, 90% of the blog is negative.  I would suggest it could be rewritten to highlight security issues all the Photobox brands, and all companies who handle PII, deal with on a daily basis, and refer more briefly to the Moonpig shambles.